BOM_AUTOMATION/requirements.txt
# ============================
# Core Numerical & Data Tools
# ============================
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1

# ============================
# Computer Vision / Image Work
# ============================
opencv-python==4.8.1.78
# If you do NOT need GUI windows like cv2.imshow, use this instead:
# opencv-python-headless==4.8.1.78

pillow==10.3.0

# ============================
# PDF Processing
# ============================
pypdf2==3.0.1
pdfplumber==0.10.3

# ============================
# Automation (Keyboard, mouse, screen)
# ============================
pyautogui==0.9.54
pynput==1.7.6

# ============================
# Web & API Communication
# ============================
requests==2.31.0
beautifulsoup4==4.12.3

# ============================
# CLI Tools / Logging / Config
# ============================
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.66.2
loguru==0.7.2

# ============================
# Utilities
# ============================
python-dateutil==2.9.0
typing-extensions==4.11.0

# ============================
# Optional: Excel Automation (COM)
# Uncomment if using Excel automation via Windows COM
# ============================
# pywin32==306
fitz
# Core Dependencies
fastapi
uvicorn
python-multipart
pymupdf
pillow
pytesseract
opencv-python-headless
numpy
pymongo
pytest

# PDF Processing
pdfplumber
pdfminer.six
ezdxf

# Computer Vision & ML
opencv-python
scikit-image
ultralytics  # YOLOv8
torch  # PyTorch for ML models
torchvision

# Feature Matching
opencv-contrib-python  # For SIFT, AKAZE

# NLP & Text Processing
spacy
python-dotenv

# Database
psycopg2-binary  # PostgreSQL
asyncpg  # Async PostgreSQL
sqlalchemy

# Data Processing
pandas
openpyxl  # Excel export

# Storage (optional)
boto3  # AWS S3
minio  # MinIO S3-compatible

# Utilities
python-dateutil
